fix: removed wrapNativeToken hook
